Removes some static_cast, make some refs const refs
Removing of static_cast has been done to make VS warns me about some precision loss and stuff : I need to manage the addition and substraction stuff better than now, like using modulos or things like this in case of overflow or underflow

+++ chip8emu/Machine.cpp
@@ -168,7 +168,7 @@ void Chip8::Machine::Execute(std::uint16_t opcode)
                 {
                     const std::int16_t sub = m_memory.VX[x] - m_memory.VX[y];
                     m_memory.VX[0xF] = (sub > 0) ? 1 : 0;
-                    m_memory.VX[x] = static_cast<uint8_t>(sub); // TODO wtf? need to check how to handle this properly
+                    m_memory.VX[x] = sub; // TODO wtf? need to check how to handle this properly
                     break;
                 }
 
@@ -181,7 +181,7 @@ void Chip8::Machine::Execute(std::uint16_t opcode)
                 {
                     const std::int16_t sub = m_memory.VX[y] - m_memory.VX[x];
                     m_memory.VX[0xF] = (sub > 0) ? 1 : 0;
-                    m_memory.VX[x] = static_cast<uint8_t>(sub); // TODO wtf? need to check how to handle this properly
+                    m_memory.VX[x] = sub; // TODO wtf? need to check how to handle this properly
                     break;
                 }
